refactoring reduce code
