+++ pkg/reduce/pnf/ordered.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package readloop
+package pnf
 
 import (
 	"container/list"
@@ -23,6 +23,7 @@ import (
 	"sync"
 	"time"
 
+	dfv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
 	"github.com/numaproj/numaflow/pkg/watermark/wmb"
 	"go.uber.org/zap"
 
@@ -32,41 +33,59 @@ import (
 	"github.com/numaproj/numaflow/pkg/reduce/applier"
 	"github.com/numaproj/numaflow/pkg/reduce/pbq"
 	"github.com/numaproj/numaflow/pkg/reduce/pbq/partition"
-	"github.com/numaproj/numaflow/pkg/reduce/pnf"
 	"github.com/numaproj/numaflow/pkg/shared/logging"
 	"github.com/numaproj/numaflow/pkg/watermark/publish"
 )
 
 var retryDelay = 1 * time.Second
 
-// task wraps the `ProcessAndForward`.
-type task struct {
-	// doneCh is used to notify when the task has been completed.
+// ForwardTask wraps the `processAndForward`.
+type ForwardTask struct {
+	// doneCh is used to notify when the ForwardTask has been completed.
 	doneCh chan struct{}
-	pf     *pnf.ProcessAndForward
+	pf     *processAndForward
 }
 
-// orderedForwarder orders the forwarding of the result of the execution of the tasks, even though the tasks itself are
+// OrderedProcessor orders the forwarding of the result of the execution of the tasks, even though the tasks itself are
 // run concurrently in an out of ordered fashion.
-type orderedForwarder struct {
+type OrderedProcessor struct {
 	vertexName    string
 	pipelineName  string
 	vertexReplica int32
 	sync.RWMutex
-	taskDone  chan struct{}
-	taskQueue *list.List
-	log       *zap.SugaredLogger
+	taskDone            chan struct{}
+	taskQueue           *list.List
+	pbqManager          *pbq.Manager
+	udf                 applier.ReduceApplier
+	toBuffers           map[string]isb.BufferWriter
+	whereToDecider      forward.ToWhichStepDecider
+	watermarkPublishers map[string]publish.Publisher
+	idleManager         *wmb.IdleManager
+	log                 *zap.SugaredLogger
 }
 
-// newOrderedForwarder returns an orderedForwarder.
-func newOrderedForwarder(ctx context.Context, vertexName string, pipelineName string, vr int32) *orderedForwarder {
-	of := &orderedForwarder{
-		vertexName:    vertexName,
-		pipelineName:  pipelineName,
-		vertexReplica: vr,
-		taskDone:      make(chan struct{}),
-		taskQueue:     list.New(),
-		log:           logging.FromContext(ctx),
+// NewOrderedProcessor returns an OrderedProcessor.
+func NewOrderedProcessor(ctx context.Context,
+	vertexInstance *dfv1.VertexInstance,
+	udf applier.ReduceApplier,
+	toBuffers map[string]isb.BufferWriter,
+	pbqManager *pbq.Manager,
+	whereToDecider forward.ToWhichStepDecider,
+	watermarkPublishers map[string]publish.Publisher,
+	idleManager *wmb.IdleManager) *OrderedProcessor {
+	of := &OrderedProcessor{
+		vertexName:          vertexInstance.Vertex.Spec.Name,
+		pipelineName:        vertexInstance.Vertex.Spec.PipelineName,
+		vertexReplica:       vertexInstance.Replica,
+		taskDone:            make(chan struct{}),
+		taskQueue:           list.New(),
+		pbqManager:          pbqManager,
+		udf:                 udf,
+		toBuffers:           toBuffers,
+		whereToDecider:      whereToDecider,
+		watermarkPublishers: watermarkPublishers,
+		idleManager:         idleManager,
+		log:                 logging.FromContext(ctx),
 	}
 
 	go of.forward(ctx)
@@ -74,26 +93,24 @@ func newOrderedForwarder(ctx context.Context, vertexName string, pipelineName st
 	return of
 }
 
-func (of *orderedForwarder) insertTask(t *task) {
+func (of *OrderedProcessor) InsertTask(t *ForwardTask) {
 	of.Lock()
 	defer of.Unlock()
 	of.taskQueue.PushBack(t)
 }
 
-// schedulePnF creates and schedules the PnF routine.
-func (of *orderedForwarder) schedulePnF(
+// SchedulePnF creates and schedules the PnF routine.
+func (of *OrderedProcessor) SchedulePnF(
 	ctx context.Context,
-	udf applier.ReduceApplier,
-	pbq pbq.Reader,
-	partitionID partition.ID,
-	toBuffers map[string]isb.BufferWriter,
-	whereToDecider forward.ToWhichStepDecider,
-	pw map[string]publish.Publisher,
-	idleManager *wmb.IdleManager) *task {
+	partitionID partition.ID) *ForwardTask {
 
-	pf := pnf.NewProcessAndForward(ctx, of.vertexName, of.pipelineName, of.vertexReplica, partitionID, udf, pbq, toBuffers, whereToDecider, pw, idleManager)
+	pbq := of.pbqManager.GetPBQ(partitionID)
+
+	pf := newProcessAndForward(ctx, of.vertexName, of.pipelineName, of.vertexReplica, partitionID,
+		of.udf, pbq, of.toBuffers, of.whereToDecider, of.watermarkPublishers, of.idleManager)
+
 	doneCh := make(chan struct{})
-	t := &task{
+	t := &ForwardTask{
 		doneCh: doneCh,
 		pf:     pf,
 	}
@@ -110,7 +127,7 @@ func (of *orderedForwarder) schedulePnF(
 
 // reduceOp invokes the reduce function. The reducer is a long-running function since we stream in the data and it has
 // to wait for the close-of-book on the PBQ to materialize the result.
-func (of *orderedForwarder) reduceOp(ctx context.Context, t *task) {
+func (of *OrderedProcessor) reduceOp(ctx context.Context, t *ForwardTask) {
 	start := time.Now()
 	for {
 		// FIXME: this error handling won't work with streams. We cannot do infinite retries
@@ -130,6 +147,7 @@ func (of *orderedForwarder) reduceOp(ctx context.Context, t *task) {
 		of.log.Errorw("Process failed", zap.String("partitionID", t.pf.PartitionID.String()), zap.Error(err))
 		time.Sleep(retryDelay)
 	}
+
 	// indicate that we are done with reduce UDF invocation.
 	close(t.doneCh)
 	of.log.Debugw("Process->Reduce call took ", zap.String("partitionID", t.pf.PartitionID.String()), zap.Int64("duration(ms)", time.Since(start).Milliseconds()))
@@ -142,11 +160,11 @@ func (of *orderedForwarder) reduceOp(ctx context.Context, t *task) {
 	}
 }
 
-// forward monitors the task queue, as soon as the task at the head of the queue has been completed, the result is
+// forward monitors the ForwardTask queue, as soon as the ForwardTask at the head of the queue has been completed, the result is
 // forwarded to the next ISB. It keeps doing this for forever or until ctx.Done() happens.
-func (of *orderedForwarder) forward(ctx context.Context) {
+func (of *OrderedProcessor) forward(ctx context.Context) {
 	var currElement *list.Element
-	var t *task
+	var t *ForwardTask
 
 	for {
 		start := time.Now()
@@ -154,16 +172,16 @@ func (of *orderedForwarder) forward(ctx context.Context) {
 		select {
 		case <-of.taskDone:
 		case <-ctx.Done():
-			of.log.Infow("Forward exiting while waiting for task completion event", zap.Error(ctx.Err()))
+			of.log.Infow("forward exiting while waiting for ForwardTask completion event", zap.Error(ctx.Err()))
 			return
 		}
 		of.log.Debugw("Time waited for a completion event to happen ", zap.Int64("duration(ms)", time.Since(start).Milliseconds()))
 
 		// a signal does not mean that we have any pending work to be done because
-		// for every signal, we try to empty out the task-queue.
+		// for every signal, we try to empty out the ForwardTask-queue.
 		of.RLock()
 		n := of.taskQueue.Len()
-		of.log.Debugw("Received a signal in task queue ", zap.Int("task count", n))
+		of.log.Debugw("Received a signal in ForwardTask queue ", zap.Int("ForwardTask count", n))
 
 		of.RUnlock()
 		// n could be 0 because we have emptied the queue
@@ -176,10 +194,10 @@ func (of *orderedForwarder) forward(ctx context.Context) {
 		currElement = of.taskQueue.Front()
 		of.RUnlock()
 
-		// empty out the entire task-queue everytime there has been some work done
+		// empty out the entire ForwardTask-queue everytime there has been some work done
 		startLoop := time.Now()
 		for i := 0; i < n; i++ {
-			t = currElement.Value.(*task)
+			t = currElement.Value.(*ForwardTask)
 			select {
 			case <-t.doneCh:
 				for {
@@ -201,10 +219,10 @@ func (of *orderedForwarder) forward(ctx context.Context) {
 					metrics.LabelVertexReplicaIndex: strconv.Itoa(int(of.vertexReplica)),
 				}).Dec()
 
-				of.log.Debugw("Removing task post forward call", zap.String("partitionID", t.pf.PartitionID.String()))
+				of.log.Debugw("Removing ForwardTask post forward call", zap.String("partitionID", t.pf.PartitionID.String()))
 				of.Unlock()
 			case <-ctx.Done():
-				of.log.Infow("Forward exiting while waiting on the head of the queue task", zap.String("partitionID", t.pf.PartitionID.String()), zap.Error(ctx.Err()))
+				of.log.Infow("forward exiting while waiting on the head of the queue ForwardTask", zap.String("partitionID", t.pf.PartitionID.String()), zap.Error(ctx.Err()))
 				return
 			}
 		}
